#1 Consertos em alguns problemas que estavam no código e uma leve melhoria na qualiade e na legibilidade também

+++ sicsec/src/main/java/br/cefetmg/sicsec/Repository/TentativaRepository.java
@@ -9,7 +9,7 @@ import br.cefetmg.sicsec.Model.Tentativa;
 public interface TentativaRepository extends JpaRepository<Tentativa, Long> {
     long countByAtividade_Id(Long atividadeId);
 
-    long countByAtividade_IdAndAbertaFalse(Long atividadeId);
+    int countByAtividade_IdAndAbertaFalse(Long atividadeId);
 
     Optional<Tentativa> findTopByAtividade_IdOrderByNumTentativaDesc(Long atividadeId);
